refactor: Component interface 타입 분리
