+++ src/components/RestaurantListHeader.ts
@@ -1,4 +1,5 @@
-import type { Component, Category, SortBy } from '../type';
+import type { Component } from '../interface';
+import type { Category, SortBy } from '../type';
 import { CATEGORIES } from '../utils/constants';
 
 type RestaurantListHeaderState = {
