+++ src/interface.ts
@@ -1,0 +1,6 @@
+export interface Component<T> {
+  $component: HTMLElement;
+  state: T;
+  setState: (state: T) => void;
+  render: () => void;
+}
