+++ src/pages/AddRestaurantDrawer/index.ts
@@ -1,4 +1,5 @@
-import type { Component, Restaurant } from '../../type';
+import type { Component } from '../../interface';
+import type { Restaurant } from '../../type';
 import {
   CATEGORIES,
   DEFAULT_CATEGORY,
