+++ src/type.ts
@@ -1,10 +1,3 @@
-export interface Component<T> {
-  $component: HTMLElement;
-  state: T;
-  setState: (state: T) => void;
-  render: () => void;
-}
-
 export type Category = '전체' | '한식' | '중식' | '일식' | '양식' | '아시안' | '기타';
 export type SortBy = 'name' | 'distance';
 export type Distance = 5 | 10 | 15 | 20 | 30;
